// src/lib/meta-types.ts
export interface AnuncioDia {
  dia: Date;
  impressoes: number;
  alcance: number;
  investimento: number;
  cliques: number;
  visualizacoesLP: number;
  frequencia: number;
  registrosCompletos: number;
  cpm: number;
  cpc: number;
  ctr: number;
  campanha: string;
  thruPlay: number;
  video25: number;
  video50: number;
  video75: number;
  anuncio: string;
  conjunto: string;
  permalink: string;
  checkouts: number;
  vendasMeta: number;
  thumbnail: string;
  status: string;
  conversasIniciadas: number;
}